[JUnit] Provide class in description for NoStepDescriptions runner
Surefire 47 uses class names to group test reports. To ensure the tests results
end up in the proper report we provide the class of the runner. We can't fake
this because the class must be loadable by surefire.

Note: This requires the junit option  `--no-step-notifications` to be enabled.

Related issues:
- https://github.com/cucumber/cucumber-jvm/issues/865